requirements.txt
django>=6.0,<7.0
pillow
gunicorn
django-markdownx
markdown
